- ajustes para o SSAO funcionar

+++ glsl/ex_47/ssao.fs
@@ -16,8 +16,8 @@ uniform int kernelSize;
 uniform float radius;
 uniform float bias;
 
-uniform int width;
-uniform int height;
+uniform float width;
+uniform float height;
 
 layout (std140) uniform Matrices {
     mat4 projection;
@@ -42,7 +42,7 @@ void main()
     {
         // get sample position
         vec3 samplePos = TBN * samples[i]; // from tangent to view-space
-        samplePos = fragPos + samplePos * radius; 
+        samplePos = fragPos + samplePos * radius;
         
         // project sample position (to sample texture) (to get position on screen/texture)
         vec4 offset = vec4(samplePos, 1.0);
